ISearchRequest serialization prototype
This commit is a minimum viable prototype of using utf8json serializer to
serialize an ISearchRequest. Utf8Json unit tests have been used to
test serialization of MatchQuery and BoolQuery.

Notable points:

1. utf8json does not have a similar concept to [JsonObject(MemberSerialization.OptIn)] to
only serialize those members that have been explicitly attributed with DataMemberAttribute.
This is something that would ideally be needed as it is cumbersome to set [IgnoreDataMember]
on all types that should be ignored.
2. ConnectionSettings is retrieved by casting IJsonFormatterResolver to a known concrete
implementation that exposes them a s a property. Not ideal, but it works.
3. utf8json does not make a distinction between an integer token and a float token as Json.NET
does. This is not much of a problem, since the bytes for the token can be inspected to determine
if they contain a decimal point, and use utf8json's internals to deserialize accordingly. Also, this
is needed only in cases where an integer/double distinction is necessary. See FuzzinessFormatter
for an example.
4. The equivalent to JsonConverter, IJsonFormatter<T> only has a generic variant. In several places
in the client, we may serialize using the an interface, but deserialize using the concrete implementation.
This is handled by ConcreteInterfaceFormatter<TConcrete, TInterface>, where the formatter
is IJsonFormatter<TInterface>. An interesting case is when the concrete type should be serialized
as the interface; in such scenarios, end up with two formatters, one for the concrete type and one
for the interface, where formatters reference the others' serialize/deserialize implementations. See
QueryContainerFormatter and QueryContainerInterfaceFormatter for examples.