Fix CrossClusterReplicationFollowTests integration test
Relates: elastic/elasticsearch#36647

This commit fixes the CrossClusterReplicationFollowTests to make assertions based on the
Elasticsearch version under test.

The follower stats API response behaviour changed from 6.5.3 to 6.5.4 - in 6.5.3 and previous, a request to e.g.

GET /<indexname>/_ccr/stats

returns in the `"indices"` array a collection of indices including `<indexname>` but also including other indices that are not `<indexname>` (and are not even a wildcard match for it).

In 6.5.4 and above, that call returns only the single index named `<indexname>` in the `"indices"` array.

(cherry picked from commit 66cddbc5432272719ffeb32de9f431a1d1f358de)