Mark internal parameterless ctors as target for deserialization
This commit marks the generated internal parameterless ctor with
SerializationConstructorAttribute, so that it is the targeted ctor for deserialization.

Without this, in the event no target best match ctor is found and the type
is determined to be side-effect free, System.Runtime.Serialization.FormatterServices.GetUninitializedObject()
is used to construct the type. For types such as CreateIndexRequest, ShrinkIndexRequest, SplitIndexRequest
e.g.

https://github.com/elastic/elasticsearch-net/blob/d20fc06904f6c194a573314a6aa6df2a5d2c742e/src/Nest/_Generated/_Requests.generated.cs#L1497

and the static initialization of RequestParameters

https://github.com/elastic/elasticsearch-net/blob/d20fc06904f6c194a573314a6aa6df2a5d2c742e/src/Nest/CommonAbstractions/Request/RequestBase.cs#L57-L66

that intialize properties with static initializers, FormatterServices.GetUninitializedObject() does not perform this static initialization,
leading to null reference exceptions when the subsequent property setters within the generated deserialization
method attempts to set a property that is relying on this static initialization.

We should consider for future in changing how utf8json determines side-effect free constructor types,
to take into account static initilization when determining whether to use FormatterServices.GetUninitializedObject()
to construct a type.