Refactor ServerError, Error and ErrorCause
With both the low level client and high level client now using the same internal serializer, the deserialization of ServerError, Error and ErrorCause can be simplified. ErrorCauseMetadata is removed from ErrorCause, and the properties it contains are hoisted onto ErrorCause. It is possible that the JSON for and error cause contains properties that are not mapped on ErrorCause. Currently, these will be ignored. If we fill it is worthwhile, we could also capture these.