Docs: Drop subtitle in Asciidoctor (#3676)
When we use Asciidoctor to build the docs for the elasticsearch-net repo
it sees the `:` in the title of the book as a delimiter for a subtitle.
This is probably a nice feature, but we actually intend the title of the
book to have a `:` in it. The simple work around is to use a different
character for the delimiter. This change does that.

(cherry picked from commit e24fdc06816758a69be270eb0e8e2f60636e1d82)