Fix utf8json escape characters
This commit fixes the logic for determining whether a double quote character is escaped when reading inside of a JSON string. the current implementation checks that the previous character is \, and that the character before this is not \, but does not take into account whether that last character itself is part is escaped. The fix is to backtrack from the double quote and count the number of sequential backslash characters; an even number implies that the double quote is not escaped and therefore the end of the JSON string.