Refactor ServerError, Error and ErrorCause (#3677)
* Refactor ServerError, Error and ErrorCause
* Include server error in integration test failure setup
* Capture additional properties in AdditonalProperties field

With both the low level client and high level client now using the same internal serializer, the deserialization of ServerError, Error and ErrorCause can be simplified. ErrorCauseMetadata is removed from ErrorCause, and the properties it contains are hoisted onto ErrorCause. It is possible that the JSON for and error cause contains properties that are not mapped on ErrorCause. These will be deserialized in key/value pairs in AdditionalProperties.