Fix Query DSL usage tests (#3804)
This commit removes the lenient assertion within QueryDslUsagesTestsBase that allowed parsing exceptions in Elasticsearch to pass. This was a temporary solution whilst upgrading the client to work with Elasticsearch 7.x.

Add an IGeoShape property to Project called LocationShape, and always seeds it with a PointGeoShape. Rename Location to LocationPoint

Update the GeoShapeFormatter to skip over the value of a field within JSON when it is not the type field. String allocations for property names are reduced by comparing fields to static byte fields.

Fix the serialized ordering of lat lon values in the geo_shape query using an envelope, since server side validation only allows valid envelopes.

Skip multi_point and circle geo_shape queries for now as both result in 400 error responses.

Do not set id when MoreLikeThis query input is a document

GeoPolygon query requires a minimum of 3 coordinates