Add User-Agent string to API calls to Elasticsearch (#3784)
This commit adds a User-Agent to all API calls to Elasticsearch, that includes the client name and version, along with OS, Framework and assembly.

Where System.Runtime.InteropServices.RuntimeInformation is available, use the properties of this type. For .NET 4.6.1 which does not support this type, use native Windows functions to determine the version of Windows, and Environment.OSVersion.VersionString for other platforms. The latter is not used for Windows because it can report incorrect version information depending on runtime version and service packs installed, and the context in which calls are executed i.e. it is not reliable.

Add static readonly strings for default user agents for both the low level and high level clients

(cherry picked from commit cb897351e22e99d8b59c6c5235245c5bdeaea718)