Fix 7.x documentation (#3821)
* run doc generation
* Prefix <example> elements in docs

This commit prefixes <example> elements in XML docs with "For example, " in generated asciidoc.
Copy the XML docs from IConnectionSettingsValues to ConnectionSettingsBase<TConnectionSettings> for now
as <inheritdoc> does not work with NuDoq for writing documentation from XML docs. Revisit how XML docs
are harvested in future to fix this limitation.