Fix/todos walkthrough (#3861)
* Remove formatting TODO

This commit removes the formatting TODO. As of now, Utf8Json serializer implementation does not support formatting; this may be addressed in a later release.

* Remove QueryContainerListFormatter

No longer needed as FiltersAggregation now uses IEnumerable<QueryContainer>

* Use string for Id on response

* Keep Type on Hit<T>

* Update TODO comments

Following team review

1. delete TODOs that are no longer relevant or will not be addressed.
2. Update TODOs that will be addressed, with a better description

* AcknowledgedResponseBase IsValid checks acknowledged field

This commit updates IsValid property for all responses that return an acknowledged field in the body. When a request is made but the response indicates that it was not acknowledged, this may be grounds for deeming the response not to be valid, which should be signalled to the user.