Handle cat API 404 responses
This commit adds handling for 404 responses for cat APIs. When a 404 is returned, NEST's HttpStatusCodeClassifier deems it to be successful, which the CatResponseBuilder tries to then deserialize the response to a collection of TCatRecord. When a 404 is returned however, the response is a JSON object with an error and status. The CatResponse<TRecord> should be deserialized from this when a 404 is returned.

Fixes #3936