Handle cat API 404 responses (#3944)
This commit adds handling for 404 responses for cat APIs. When a 404 is returned, NEST's HttpStatusCodeClassifier deems it to be successful, which the CatResponseBuilder tries to then deserialize the response to a collection of TCatRecord. When a 404 is returned however, the response is a JSON object with an error and status. The CatResponse<TRecord> should be deserialized from this when a 404 is returned.

Fixes #3936

(cherry picked from commit 11a2ab3e08a28bcbff342022a6bf4bcf03ad3b0d)