Honour IPropertyMappingProvider on custom IElastisearchSerializer (#3942)
This commit fixes a bug introduced with Diagnostics Source support. DiagnosticsSerializerProxy does not implement IPropertyMappingProvider, so if it wraps and delegates to a serializer that does implement IPropertyMappingProvider, such as JsonNetSerialzer, this is not honoured when assigning to _propertyMappingProvider. This commit uses the sourceSerializer when determining _propertyMappingProvider, and not the DiagnosticsSerializerProxy.

Fixes #3926