Expose Response on GetTaskResponse (#3966)
* Expose Response on GetTaskResponse

This commit exposes Response on GetTaskResponse. Response is the response object for the action performed by the Task, so will vary by action. For this reason, expose Response as an internal LazyDocument, with a method to retrieve a TResponse from the deserialized LazyDocument bytes.

* Deserialize response with RequestResponseSerializer

This commit updates the GetResponse<T>() impl on GetTaskResponse to use the RequestResponseDeserializer when deserializing the LazyDocument. This is exposed as an internal method only.
Remove LazyDocument ctor overloads that are superfluous.

Closes #3963