Fix 6.x query validity in integration tests (#3999)
* Send queries to cluster when running integration tests

* Do not set _id and routing automatically (#3805)

This commit updates LikeDocument so that _id and routing are not
set when a document input is used.

Routing is required in the MoreLikeThisFullDocumentQueryUsageTests because the Project type mapping
requires routing, but this should not be the default.

Fixes #3977

* Fix Query DSL usage tests (#3804)

This commit removes the lenient assertion within QueryDslUsagesTestsBase that allowed parsing exceptions in Elasticsearch to pass. This was a temporary solution whilst upgrading the client to work with Elasticsearch 7.x.

Add an IGeoShape property to Project called LocationShape, and always seeds it with a PointGeoShape. Rename Location to LocationPoint

Update the GeoShapeFormatter to skip over the value of a field within JSON when it is not the type field. String allocations for property names are reduced by comparing fields to static byte fields.

Fix the serialized ordering of lat lon values in the geo_shape query using an envelope, since server side validation only allows valid envelopes.

Skip multi_point and circle geo_shape queries for now as both result in 400 error responses.

Do not set id when MoreLikeThis query input is a document

GeoPolygon query requires a minimum of 3 coordinates

* Fix span query integration tests

This commit fixes the span query integration tests. Fix namespaces.

* Fix HasChild query integration tests

This commit introduces a TypeRelation property of type RelationName and deprecates Type property. The property should be RelationName to correctly infer the relation.

* Valid script in script query usage tests

* Add routing value to More like this query

* Use tag nested field in nested query usage tests

* skip geo shape query tests on >=6.6.0

This commit skips geo shape queries for circles and multi_point above 6.6.0. These queries are not supported against BKD backed shapes

* fix query_string query integration tests

* Update docs

* Assert AfterKey is not null only on >= 6.3.0

* Use static First Project property

* Valid queries