Serialize GeoLocation with internal serializer (#3997)
This commit adds GeoLocation to types handled by the internal serializer so that
a property on a document POCO of type GeoLocation is serialized by the internal serializer.

It's not possible with the current test configuration to test this scenario, because
IL rewriting happens after unit tests are run, so the Json.NET constructs used on Nest types
are recognised by the JsonNetSerializer. Have added test that _would_ test behaviour, should
tests be run after IL-merging in the future.

Fixes #3981

(cherry picked from commit 623c724a3205b21ffca9cf88338f94697efef677)