Generate client doc examples from Elasticsearch reference examples (#3993)
Relates: elastic/clients-team#49, #3993

This commit adds two new projects to the solution for implementing .NET client examples for the examples in the Elasticsearch reference documentation:

- Examples
A class library of examples classes with methods to represent the C# implementations of Elasticsearch reference doc examples

- ExamplesGenerator

An executable that can generate C# example classes from an asciidoc reference file. 

Source reference examples are collected into a reference page, and a C# *Page class generated for the reference page, in a directory that matches the source asciidoc relative path, in the Examples project. Each reference example is generated as a method. If a *Page class already exists, is is cloned, then

1. any new examples that match the hash of an old example will be added. The example method name will be updated, if necessary
2. any new examples that do not already exist will be added
3. any old examples that do not exist in the new examples will be removed

- Allow supplying a delegate to MatchesExample to be able to alter the reference example when it uses constructs that the client does not support. For example

1. _source=*.id as a shorthand for _source_includes=*.id
2. short form of certain queries in query DSL like match query { "match": { "title":  "Search"  }}