Ensure GeoLocationFormatter returns GeoLocation type (#4026)
The ReindexApiTests were failing on a serialisation roundtrip when dealing with properties that used the GeoLocation type. This manifested due to the inclusion of GeoLocation types in the _source serialisation on commit https://github.com/elastic/elasticsearch-net/commit/5492599e7d5964bbc0c9f7975f81ad5db276ea4e

The GeoLocationFormatter was found to be returning a GeoCoordinate type and not a GeoLocation type.