Serialize first precision value on GeoSuggest
Relates: #4091

This commit adds a SerializeAsSingleFormatter<T> to serialize only the first value in the Precision collection
for a GeoSuggestContext context mapping for the "completion" field data type.

Move the Context mapping related types to the Nest/Mapping/Types/Specialized/Completion directory to disambiguate
from the context query types.

Deprecate the Neighbors property on IGeoSuggestContext.