Better randomized initial nodes test
This commit adds a better randomized initial nodes test by introducing a Thread.Sleep of 1 millisecond between successive instantiations of StaticConnectionPool. A new instance of Random is created for each instance of StaticConnectionPool, and the Random is initialized with Environment.TickCount, the number of milliseconds since the system started. To therefore ensure that seeds for successive instances of Random are different, a wait is introduced.