Update REST API specs to v7.4.0 (#4117)
* Transform REST API spec to existing format

This commit changes the in memory structure of the ApiEndpoint in 7.4.0 to the old structure expected by the ApiGenerator.

* Update specs to v7.4.0

This commit updates the REST API specs to those in the v7.4.0 branch, and re-runs the ApiGenerator. New 7.4.0 APIs are excluded from code generation for now.

* Patch 7.4.0 breaking changes

- delete_by_query
- msearch_template
- search_template

(cherry picked from commit 6da223b4539c6099495fa3113e84314760c4c690)