Implement Snapshot Lifecycle Management APIs
This commit implements the Snapshot Lifecycle APIs in both the high level and low level clients.
Unit tests for URLs, integration tests for the SLM API flow.

The slm.get_lifecycle JSON spec is patched so that policy_id is a list and the Ids type is used to model multiple ids.