Serialize dates with min 3 fractional digits when using SourceSerializer in integration tests
This commit adds a new JsonConverter to serialize DateTime/DateTimeOffset with a minimum of 3 fractional digits when running integration tests using the test SourceSerializer. Related to a bug in Elasticsearch <7.1.0 in the XML comment of the JsonConverter.

Always emit the random values used when running a test/integrate build, not just if they were explicitly set.