Correctly serialize _geo_distance sort
This commit updates the SortFormatter to emit a value separator when serializing a _geo_distance only when
the IGeoDistanceSort contains properties with values. This is determined by checking the serialized bytes
are more than 2 i.e. an empty JSON object, {}.

Add additional _geo_distance sort to SortUsage tests to cover when IGeoDistanceSort contains/doesn't contain properties with
values.