Fix #4289 low level client index http method generation wrong (#4290)
* Fix #4289 low level client index http method generation wrong

We used to always send POST which is not right either but the server
accepted it. Now that we have a new format and transform to the old
format in the interim we only picked up PUT for all which is not correct
for index operations without an id.

* add unit tests

(cherry picked from commit 02d40766cf4fcdd660ca1af0ae818bb8efb57bc3)