Read numeric aggs keys as long or double
This commit updates the reading of numeric aggregation keys to read as double only when the byte representation of the value contains a decimal point, and to read it as a long otherwise. This makes the assumption that whole numbers returned are smaller than long.MaxValue. This same assumption is made in PrimitiveObjectFormatter.

Closes #4285