Read numeric aggs keys as long or double (#4310)
This commit adds an extension method to check if an ArraySegment<byte> can be converted to a long value, and updates the reading of numeric aggregation keys to read as long when the byte representation of the value is a long, and to read it as a double otherwise. Similarly, PrimitiveObjectFormatter's number handling is also updated.

Closes #4285

(cherry picked from commit cbd1b328d565e9ad9eae02a49057af62cc791a35)