Return distinct items from GetMany and SourceMany
This commit fixes a bug where a GetMany or SourceMany API call
with a repeated id would return a cartesian product of id and documents.

Take for example the same id repeated 3 times; Elasticsearch will return
3 documents in the response, where each JSON object is a representation
of the same underlying document. Since all 3 documents have the same id,
GetMany and SourceMany would return all 3 documents as a match for the first id,
all 3 as a match for the second id, and so on.

With this fix, For each id, only the first matching document is returned.
One _could_ argue that only a single document should be returned for the same
repeated id. This fix however tries to reflect what is in the
Elasticsearch response.

Fixes #4342