Use correct IJsonFormatter<T> in attributes
This commit fixes a bug where VerifyRepositoryResponse was attributed
with the wrong formatter type in JsonFormatterAttribute applied at the type level.
Utf8Json's formatters are strongly typed for speed at the cost of less flexibility.

Fix other cases where the incorrect formatter type is used, and introduce a
convention test to assert that types and their properties attributed with
JsonFormatterAttribute use the correct formatter type.

Fixes #4382