Create exception formatters for different exceptions (#4378)
This commit changes the ExceptionFormatterResolver to return a new instance
of the now generic ExceptionFormatter<TException> for the type of exception.
This fixes a bug where an ExceptionFormatter of type IJsonFormatter<Exception>
is returned for derived exception types, resulting in an InvalidCastException
at runtime.

Formatter instances created are cached in ElasticsearchNetFormatterResolver
for the type T, so will be reused once created.

Fixes #4294

(cherry picked from commit 8156ea741baf6b100f4beb724b4300e8318355ed)