Track indices seen for each id
This commit updates the GetMany and SourceMany implementations
to keep track of the seen indices for each id, to return a single
document of id and index combination _per_ id i.e. if the same id
is specified multiple times in the GetMany or SourceMany call, and
only a single index is targeted, each id will return just a single
document.