Use XPackCluster for PutPipeline API tests (#4429)
This commit changes the PutPipelineApiTests to use
an X-Pack cluster with security enabled. As part of this change,
the ingest attachment plugin must also be installed. This is deemed a
better compromise than specifying a new cluster type for this test only.

This change also fixes a NPE thrown by set_security_user processor, which
appears to happen when security is disabled. See elastic/elasticsearch#52474

Closes #4385