Fix yaml tests not deleting newly defined hidden indices
Before we were not deleting system indices only `.ml-*`

However this started to become an issue because it left closed indices
around. This might be related to .ml indices now being hidden.

Warrants more investigation as the new solution deleting all indices all
the time is definitely a bit slower