Fix GetMany when no default index is set (#4463)
https://github.com/elastic/elasticsearch-net/pull/4353

Fixed an issue with the GetMany helpers that returned the cartesian product of all ids specified rather
then creating a distinct list if more then one index was targeted.

This PR also updated the routine in the serializer to omit the index name from each item if the index is
already specified on the url in case of multiple indices

This updated routine in the `7.6.0` could throw if you are calling:

    client.GetMany<T>(ids, "indexName");

Without configuring `ConnectionSettings()` with either a default index for T or a global default index.

(cherry picked from commit 87c8cdd05ba3ef3818e6cd3444bd71b5de84d443)