Deduplicate example method names (#4485)
* Deduplicate example method names

This commit deduplicates the generated example method names,
when generating C# methods from the examples. The method names
follow the convention of Line<linenum>, where linenum is the
line number in the alternative report. The alternative report
can report duplicate line numbers however, as reported in

elastic/docs#1760

so deduplicate in the generator.

* update the Description Attribute, ensuring
it stays in sync with the source.
* log out duplicate example line numbers