Don't sanitize the keys for Filters aggregation named buckets (#4656)
This commit fixes a bug where the keys in the dictionary passed
to filters aggregation are sanitized when copying to the backing
dictionary. Keys in this scenario are the keys for named buckets
which should never be sanitized.

Fixes #4582

(cherry picked from commit 7d522feaf00fdbee0ac20961f3c7b38a502d0260)