Fix check for agressive connectionlimit (#4663)
* Fix check for agressive connectionlimit

In https://github.com/dotnet/runtime/issues/22366 we found that if
HttpClient is using the curl handler it will lead to TCP connections
bleeding. Our DefaultConnectionLimit is very restrictive if this is
true. Our check however is too lenient and will default to true always
on .NET core since netcoreapp still ships with CurlHandler as recent as
`3.1.x`

* remove unused reference

* Detect if CurlHandler exists

This commit updates the implementation of determining if
a conservative default connection limit should be set by also checking
if CurlHandler exists when UseSocketsHttpHandler exists and is
disabled.

Co-authored-by: Russ Cam <russ.cam@forloop.co.uk>