[Backport 6x] Fix check for agressive connectionlimit (#4669)
In dotnet/runtime#22366 we found that if
`HttpClient` is using the curl handler it will lead to TCP connections
bleeding. Our DefaultConnectionLimit is very restrictive if this is
true. Our check however is too lenient and will default to true always
on .NET core since netcoreapp still ships with CurlHandler as recent as
`3.1.x`