Fix api url lookups (#4714)
* Exclude lookups for ignored high level APIs

This commit fixes a bug in the ApiUrlsLookup generator
to use the endpoint filename when comparing against the list
of high level APIs to ignore.

* Do not remove Index from SimulateIndexTemplate method name

This commit fixes a bug in the CSharp method name generation
where Index is removed from the name of the SimulateIndexTemplate
method in the Indices namespace. Removing this causes a collision
with SimulateTemplate

(cherry picked from commit d36aabbd999ce0029d530c368839217fc939d0d7)