Introduce .ci/make.sh
Initial stab at common entry script for all the Elasticsearch clients.

This puts our build through a docker container so that folks/CI does not need
to install dependencies on their machines to build artifacts.

> ./.ci/make.sh release VERSION

will output our packages, release notes, API reports in `.ci/output`.

This also updates our build to use: https://github.com/nullean/nupkg-validator

This is the tool version of our build script targets that

- validate the dlls are build in release mode
- validate the dll version numbers
- validate the dll assemly identity
- validate the dll assembly names.

With the added benefit that the dependency on `sn` is gone which is not
available on the dotnet base docker images that we use.

`.ci/make.sh` is not done yet, will iterate on this as the flow in the
unified release manager becomes more apparant. For example right now we
bump the version after release but we might need to do so prior.