Consistent file string comparison across Windows and Linux
This commit fixes a difference in the default string comparison
on .NET core across Windows and Linux which was causing the
Stale docs github action to fail on CI because generated documentation
was ordering asciidoc includes and list items in a different order.

This fix sets a specific culture to use and uses ordinal string
comparison.