Enable TCP states and ThreadPool stats
This commit adds the ability to enable TCP and ThreadPool
statistics to be collected when making a request, exposing
these statistics on the IApiCallDetails on the response, as
well as including in the debug information.

Add documentation explaining how to include these statistics
either on a global or per request basis.