Remove use of TypeInfo
This commit removes the use of TypeInfo from the projects.
TypeInfo was introduced when full reflection was not available
in .NET Core and the libraries supported NetStandard 1.3.

Now that the minimum TFMs supported are .NET 4.6.1 and
.NET Standard 2.0, there's no need to use TypeInfo.

(cherry picked from commit 92e81a90d40c1ce2489618ec35c72592ba75f8d2)