Align serialization and property walker GetProperties() implementations
This commit aligns the implementation for retrieving properties when
determining property mapping for a type with the way that properties
are retrieved when generating a serialization formatter for a type,
in relation to how member hiding and overridden members are determined.

Replace usages of GetGetMethod() with the GetMethod property and
GetSetMethod() with the SetMethod property.

(cherry picked from commit 0e4e99d636b838edb434584330afdad14fcbc7ad)