Add ExceptionConverter, do not serialize nulls
This commit adds an ExceptionConverter to serialize
exceptions with System.Text.Json in the same manner
as Utf8Json.

Do not serialize nulls with System.Text.Json, and
allow a derived serializer to change JsonSerializerOptions.
JsonSerializerOptions are lazily constructed in the ctor
because they are created by a virtual method.