Introduce assemble of SNAPSHOT builds
Without affecting the version number stored in global.json

`.ci/make.sh` is the common entry for all clients builds and one of its
jobs is to build SNAPSHOT builds.

Long term we want to move to minver which means we no longer need to
store the version in global.json.